requirements.txt
asgiref==3.7.2
Babel==2.12.1
Django==4.2.5
django-crispy-forms==2.0
django-money==3.2.0
py-moneyed==3.0
sqlparse==0.4.4
typing_extensions==4.7.1
